OPENNLP-1894: Apply the review-convention pass and drop unreferenced lexicon accessors
